Read EventQRCode & Check VoteQRCode_eventId OK
